requirements.txt
pytest==8.2.1
Appium-Python-Client==4.0.0
selenium==4.21.0